requirements.txt
anthropic>=0.44.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-telegram-bot>=20.7
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
psutil>=5.9.0
aiofiles>=23.0
python-multipart>=0.0.9
